ask-my-book/ask_my_doc/backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
# cognee needs multipart>=0.0.22; fastapi-users<15.0.5 pins 0.0.21 → force 15.0.5+
python-multipart>=0.0.22,<0.1.0
fastapi-users>=15.0.5
cognee==0.5.3
langchain==1.2.13
langchain-openai==1.1.12
langchain-community==0.4.1
langgraph==1.1.3
openai==2.29.0
python-dotenv==1.2.2
pypdf==6.9.2
tiktoken==0.12.0
PyPDF2==3.0.1
mistralai==1.12.4
instructor==1.14.5
psycopg2-binary==2.9.11
bcrypt==4.0.1 
passlib==1.7.4
python-jose[cryptography]==3.5.0
